--- hibernate-core/src/test/java/org/hibernate/test/bytecode/enhancement/lazy/group/MoreSpecializedKey.java
@@ -1,17 +1,0 @@
-/*
- * Hibernate, Relational Persistence for Idiomatic Java
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
- */
-package org.hibernate.test.bytecode.enhancement.lazy.group;
-
-import java.io.Serializable;
-import javax.persistence.Entity;
-import javax.persistence.Inheritance;
-import javax.persistence.InheritanceType;
-
-@Entity(name="MoreSpecializedKey")
-@Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
-public class MoreSpecializedKey extends SpecializedKey implements Serializable {
-}
