+++ hibernate-core/src/test/java/org/hibernate/test/bytecode/enhancement/lazy/proxy/ModelEntity.java
@@ -1,10 +1,17 @@
 /*
  * Hibernate, Relational Persistence for Idiomatic Java
  *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+
+/*
+ * Hibernate, Relational Persistence for Idiomatic Java
+ *
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.test.bytecode.enhancement.lazy.group;
+package org.hibernate.test.bytecode.enhancement.lazy.proxy;
 
 import java.sql.Timestamp;
 import javax.persistence.Basic;
