+++ hibernate-core/src/test/java/org/hibernate/test/bytecode/enhancement/lazy/proxy/MoreSpecializedKey.java
@@ -1,20 +1,24 @@
 /*
  * Hibernate, Relational Persistence for Idiomatic Java
  *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+
+/*
+ * Hibernate, Relational Persistence for Idiomatic Java
+ *
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.test.bytecode.enhancement.lazy.group;
+package org.hibernate.test.bytecode.enhancement.lazy.proxy;
 
 import java.io.Serializable;
 import javax.persistence.Entity;
 import javax.persistence.Inheritance;
 import javax.persistence.InheritanceType;
-import javax.persistence.Table;
 
-@Entity(name="GenericKey")
+@Entity(name="MoreSpecializedKey")
 @Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
-@Table(name="PP_GenericDCKey")
-public abstract class GenericKey extends AbstractKey implements Serializable {
-
+public class MoreSpecializedKey extends SpecializedKey implements Serializable {
 }
